backend/requirements.txt
django>=5.0,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-apscheduler>=0.6.2
python-dotenv>=1.0.1
feedparser>=6.0.11
trafilatura>=1.8.1
google-generativeai>=0.4.0
groq>=0.5.0
openai>=1.14.0
httpx>=0.27.0
numpy
